refactor: restructure @buntal packages by moving HTTP and middleware functionalities to separate packages
- Deleted core HTTP functionalities and moved them to @buntal/http package.
- Created a new @buntal/middlewares package to handle authentication, CORS, and logging middlewares.
- Implemented cookie handling in the new @buntal/http package.
- Updated request and response handling to align with the new package structure.
- Removed unused files and dependencies from the core package.
