--- packages/@buntal/core/middlewares/index.ts
@@ -1,3 +1,0 @@
-export { auth } from './auth'
-export { cors } from './cors'
-export { logger } from './logger'
