--- packages/@buntal/core/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@buntal/core",
-  "description": "HTTP client framework for Bun",
-  "version": "0.0.7",
-  "type": "module",
-  "homepage": "https://buntaljs.org",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mgilangjanuar/buntal.git"
-  },
-  "devDependencies": {
-    "@types/bun": "latest"
-  },
-  "peerDependencies": {
-    "typescript": "^5"
-  },
-  "dependencies": {
-    "bcryptjs": "^3.0.2",
-    "jose": "^6.0.11"
-  }
-}
