--- packages/@buntal/core/security/hash.ts
@@ -1,9 +1,0 @@
-import { compareSync, genSaltSync, hashSync } from 'bcryptjs'
-
-export function hash(password: string) {
-  return hashSync(password, genSaltSync())
-}
-
-hash.compare = function (password: string, hashed: string) {
-  return compareSync(password, hashed)
-}
