+++ packages/@buntal/middlewares/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@buntal/middlewares",
+  "version": "0.0.4",
+  "module": "index.ts",
+  "type": "module",
+  "devDependencies": {
+    "@types/bun": "latest"
+  },
+  "peerDependencies": {
+    "typescript": "^5"
+  },
+  "dependencies": {
+    "@buntal/http": "latest",
+    "jose": "^6.0.11"
+  }
+}
